Enable provisioning bastion machine through foreman
This is useful if your bastion RHEL version is below requirements
or if you want a clean bastion. The hammercli commands are run from a container
which is built on the localhost since we cannot publish an image publicly with the
foreman url.

Commit also moved the RHEL version check for ipv6 to the bastion node instead of localhost
as the playbook is now expected to be run from outside bastion.

Also handles the case where bastion is powered off and user tries to run the playbook.

Handles copying keys to bastion

Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>