Introduce `sno_node_count` variable
This helps deploy only on subset of hosts in the allocation.
Also introduces additional checks to ensure that requested
worker/sno nodes can be accomodated in the lab allocation.
If `worker_node_count` or `sno_node_count` are empty, we deploy
on all avaialable nodes after accounting for bastion and controlplane.

This also helps enforce the usage of `worker_node_count` in tasks where
it was previously not used.

Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>