Update JetLag's assisted-installer version
I'm open to suggestions about the assisted-installer version(s). I was unable
to locate any definitive release/support version information, which is a bit
frustrating. Previously the UI container was versioned differently: although
I'm using the same version currently for both, I created two separate vars to
allow each to be set separately if desired.

Updating beyond v2.29 required a change in the cluster JSON payload from
`api_vip` as a string to `api_vips` as a list of objects containing an `"id"`
field. While the current workflow passes a network interface name to create
the cluster, we don't have an IP as this point and the field is not required:
I simply dropped `api_vip` and `ingress_vip` from the "Create cluster" step
and constructed the proper object list for the "Patch cluster ingress/api vip
address" step.

I've used this to create both bare metal and SNO clusters for the latest patch
of each OpenShift release from 4.12.53 through 4.15.3 *and* the 4.16.0-ec.4
developer preview. I didn't do a lot with the constructed clusters, but the
assisted installer GUI shows a healthy cluster and I did a few simple `oc`
commands from the bastion which seem to suggest all is fine.

Resolves #458