allow user to define controlplane_network_gateway
When public vlan is true, user can't override the value for
controlplane_network_gateway. In prow, we want to use the bastion
host as the default gateway instead of vlan's default gateway
for some workloads. This allows user to specify the default gateway
for both the public and private networks. If the user not defines
it, then jetlag defines it using the existing mechanism.

Signed-off-by: venkataanil <vkommadi@redhat.com>