Allow setting the amount of hypervisors for VMNO (#741)
Introduce a variable for VMNO inventory which allows
the user to pass how many nodes in the inventory should
be used as hypervisors. This is useful when we don't
need all the nodes in the allocation to be used as
hypervisors.

The current behavior to use all the nodes in the allocation
is maintained as the default.

Signed-off-by: Marius Cornea <mcornea@redhat.com>